Fixed bias on/off switch
